fix python3 issues

+++ kapre/backend.py
@@ -169,7 +169,7 @@ def get_stft_kernels(n_dft, keras_ver='new'):
     assert n_dft > 1 and ((n_dft & (n_dft - 1)) == 0), \
         ('n_dft should be > 1 and power of 2, but n_dft == %d' % n_dft)
 
-    nb_filter = n_dft / 2 + 1
+    nb_filter = int(n_dft / 2 + 1)
     dtype = K.floatx()
 
     # prepare DFT filters
